Move session and loader tests to owners
