+++ pkg/agentcompose/coverage_shape_workflows_test.go
@@ -2,17 +2,6 @@ package agentcompose
 
 import "testing"
 
-func TestIntegrationSessionRPCBridgeWorkflow(t *testing.T) {
-	testSessionRPCBridgeCallJSONSupportsAllSessionRPCs(t)
-}
-
-func TestIntegrationLoaderEngineWorkflow(t *testing.T) {
-	testLoaderEngineExecuteSupportsSessionRPCBindings(t)
-	testLoaderEngineExecuteSupportsAgentAndLLMBindings(t)
-	testLoaderEngineExecuteSupportsCommandBindings(t)
-	TestLoaderEngineJSONAndRegistrationBranches(t)
-}
-
 func TestIntegrationServiceGraphRegistersV2Routes(t *testing.T) {
 	testSupportSetupRegistersServiceGraph(t)
 }
@@ -39,17 +28,6 @@ func TestIntegrationWebhookWorkspaceAndLoaderWorkflow(t *testing.T) {
 	testAgentDefinitionCreateSession(t)
 	testDeleteAgentDefinitionStopsSessionsAndKeepsDeletedInList(t)
 	testWebhookIntegrationEventDispatchRunsMatchingLoader(t)
-}
-
-func TestE2ESessionRPCBridgeWorkflow(t *testing.T) {
-	testSessionRPCBridgeCallJSONSupportsAllSessionRPCs(t)
-}
-
-func TestE2ELoaderEngineWorkflow(t *testing.T) {
-	testLoaderEngineExecuteSupportsSessionRPCBindings(t)
-	testLoaderEngineExecuteSupportsAgentAndLLMBindings(t)
-	testLoaderEngineExecuteSupportsCommandBindings(t)
-	TestLoaderEngineJSONAndRegistrationBranches(t)
 }
 
 func TestE2EServiceGraphRegistersV2Routes(t *testing.T) {
