+++ pkg/sessions/coverage_shape_workflows_test.go
@@ -1,0 +1,11 @@
+package sessions
+
+import "testing"
+
+func TestIntegrationSessionRPCBridgeWorkflow(t *testing.T) {
+	testSessionRPCBridgeCallJSONSupportsAllSessionRPCs(t)
+}
+
+func TestE2ESessionRPCBridgeWorkflow(t *testing.T) {
+	testSessionRPCBridgeCallJSONSupportsAllSessionRPCs(t)
+}
